[CFX-1923] Adding retraining Settings to the deployment (#144)
### retrainingSettings
It was decided to include this in the deployment_resource as it is part
of the deployment settings rather than the retraining policy. Updating
these settings should be performed on the active deployment, which is
why it was moved to a separate function.
 
### Tests
- Notebook_resources - fixed fail because of the wrong setting of the
use_case_id
- refactored obtaining os environments - based on .env file or from the
system.
 - refactored folder creation in some tests
 
### References 
[Mitigation
Retraining](https://docs.datarobot.com/en/docs/api/reference/public-api/mitigation_retraining.html#get-apiv2deploymentsdeploymentidretrainingsettings)