[-] Attempt fixing resource upload and removal (#152)
This pull request introduces batching for file uploads and file
deletions to avoid hitting API limits.

Files to be uploaded `localFiles` are divided into batches of 100 files
each.
Each batch is processed separately. This ensures that no more than 100
files are uploaded in a single API call.

Similarly, files to be deleted are processed in batches of 100 files.